Accept zero-prefixed chunk lengths
Our Transfer-Encoding: chunked implementation was not actually correct,
and as a consequence was not robust to servers that improperly add
leading zeroes to chunk lengths.  This is fixed here by more closely
conforming to the RFC spec ourselves, and removing a few parse states
that don't match the BNF defined there.

Fixes #23.